requirements.txt
djangorestframework
patterns